requirements.txt
deepgram-sdk
requests
youtube_dl
ffmpeg-python